# Conflicts:
#	jmp/mp-service/src/main/resources/db/shop/shop_update.sql

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/integration/GroupIntegrationMaVo.java
@@ -21,7 +21,7 @@ public class GroupIntegrationMaVo {
 	private Timestamp startTime;
 	private Timestamp endTime;
 	private Integer integration;
-	private Short inviteNum;
+	private Integer inviteNum;
 	private Integer inviteUser;
 	private Byte isNew;
 	private Byte isLook;
